Finall fix

+++ include/menuLab_1.h
@@ -8,14 +8,13 @@ void info();
 void inputEl();
 void solution1();
 void outputEl();
-void clearArr(double* );
+void clearArr(double *);
 
 void inpFile();
 void outFile();
 
-void outTableGraf(double*, int, int, std::string, std::string, HDC, size_t);
+void outTableGraf(double *, int, int, std::string, std::string, HDC, size_t);
 void Text(HDC hdc, int indentX, int indentY, std::string title, COLORREF textColor, COLORREF BG, int size);
-
 
 int setSizeArr();
 
@@ -23,7 +22,7 @@ std::string getNameFile();
 void replaceElFile();
 bool is_number(std::string);
 
-extern int sizeArr;
-extern double* arr;
+extern size_t sizeArr;
+extern double *arr;
 extern HDC hdc;
 extern char separate;
